Skipping the module looks better in the docs, so move it
